Instantiate calculator.java file and also added .gitignore to our project

+++ calculator.java
@@ -4,7 +4,7 @@ public class{
 		return a + b;
 		}
 
-	public double addition(int a, double b){
+	public double addition(double a, double b){
 		return a + b;
 	}
